DASH_CLUSTER_KPIS/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
openpyxl
